% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_abundance_diagrams.R
\name{abundance_diagram}
\alias{abundance_diagram}
\title{Print two abudance diagrams}
\usage{
abundance_diagram(data, site)
}
\arguments{
\item{data}{a dataframe}

\item{site}{a character chain}
}
\value{
two barplots
}
\description{
Print two abudance diagrams
}
