% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{compute_growth_BCI}
\alias{compute_growth_BCI}
\title{Compute growth for BCI}
\usage{
compute_growth_BCI(data)
}
\arguments{
\item{data}{}
}
\value{
the same dataframe with some new columns, including annual growth
}
\description{
Compute growth for BCI
}
