% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{compute_growth_paracou}
\alias{compute_growth_paracou}
\title{Compute the annual growth of each tree of the paracou dataset}
\usage{
compute_growth_paracou(Paracou)
}
\arguments{
\item{Plots}{}
}
\value{
a dataframe containing the annual growth in mm/y
}
\description{
Compute the annual growth of each tree of the paracou dataset
}
