% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{compute_growth_uppangala}
\alias{compute_growth_uppangala}
\title{Growth computation for Uppangala}
\usage{
compute_growth_uppangala(data)
}
\arguments{
\item{data}{}
}
\value{
data but with new colums
}
\description{
Growth computation for Uppangala
}
