% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{compute_mean_growth_BCI}
\alias{compute_mean_growth_BCI}
\title{Compute mean growth for BCI}
\usage{
compute_mean_growth_BCI(data)
}
\arguments{
\item{data}{}
}
\value{
a dataframe with mean individual growth
}
\description{
Compute mean growth for BCI
}
