% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{compute_mean_growth_paracou}
\alias{compute_mean_growth_paracou}
\title{Compute the mean growth for paracou}
\usage{
compute_mean_growth_paracou(Data_no_mean)
}
\arguments{
\item{Data_no_mean}{}
}
\value{
a dataframe with the mean growth of each tree
}
\description{
Compute the mean growth for paracou
}
