% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{compute_mean_growth_uppangala}
\alias{compute_mean_growth_uppangala}
\title{Compute mean growth for Uppangala}
\usage{
compute_mean_growth_uppangala(data)
}
\arguments{
\item{data}{}
}
\value{
data but with a new colums Mean_G_Tree
}
\description{
Compute mean growth for Uppangala
}
