% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{create_identspecies}
\alias{create_identspecies}
\title{Add IdentSpecies column}
\usage{
create_identspecies(data)
}
\arguments{
\item{data}{}
}
\value{
the dataframe with a new column
}
\description{
Add IdentSpecies column
}
