% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Moran.R
\name{moran_analysis}
\alias{moran_analysis}
\title{Do the Moran analysis of the data}
\usage{
moran_analysis(data)
}
\arguments{
\item{data}{}
}
\value{
a dataframe containing the results of the Moran analysis
}
\description{
Do the Moran analysis of the data
}
