% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Moran.R
\name{moran_I}
\alias{moran_I}
\title{Moran's I computation}
\usage{
moran_I(data)
}
\arguments{
\item{data}{}
}
\value{
A vector of length 4 corresponding to Moran's index, the p-value, the expected value under H0 and the standard deviation of the normal distribution
}
\description{
Moran's I computation
}
