% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_variance_comparison.R
\name{props_ind_comp_var}
\alias{props_ind_comp_var}
\title{Calculate the proportion of individuals in each category}
\usage{
props_ind_comp_var(data)
}
\arguments{
\item{data}{}
}
\value{
a vector containing the 4 proportions
}
\description{
Calculate the proportion of individuals in each category
}
