% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{read_data_bci}
\alias{read_data_bci}
\title{Load BCI data and gather all censuses in the same file}
\usage{
read_data_bci()
}
\value{
the BCI dataset used in the analysis
}
\description{
Load BCI data and gather all censuses in the same file
}
