% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{read_data_paracou}
\alias{read_data_paracou}
\title{Read the original paracou data}
\usage{
read_data_paracou()
}
\value{
a dataframe containing Paracou's data with plots merged
}
\description{
Read the original paracou data
}
