% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{rename_indet_paracou}
\alias{rename_indet_paracou}
\title{Rename the indeterminate species in paracou}
\usage{
rename_indet_paracou(Paracou_no_mean)
}
\arguments{
\item{Paracou_no_mean}{}
}
\value{
the dataframe with a modified Sp column
}
\description{
Rename the indeterminate species in paracou
}
