% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Moran.R
\name{results_moran}
\alias{results_moran}
\title{Compute the proportion of species and individuals in the different categories of results of the Moran test}
\usage{
results_moran(data1, data2, data3)
}
\arguments{
\item{data1}{a dataframe}

\item{data2}{a dataframe}

\item{data3}{a dataframe}
}
\value{
Results_moran, a dataframe
}
\description{
Compute the proportion of species and individuals in the different categories of results of the Moran test
}
