% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_variance_comparison.R
\name{results_var_comp}
\alias{results_var_comp}
\title{compute a table with the proportion of individuals and species with var inter < var intra for each dataset}
\usage{
results_var_comp(data1, data2, data3, site1, site2, site3)
}
\arguments{
\item{data1}{}

\item{data2}{}

\item{data3}{}

\item{site1}{}

\item{site2}{}

\item{site3}{}
}
\value{
a dataframe
}
\description{
compute a table with the proportion of individuals and species with var inter < var intra for each dataset
}
