% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_data_manip.R
\name{transform_Indet_species}
\alias{transform_Indet_species}
\title{Transforms the name of the undertermined species into "Indet_sp"}
\usage{
transform_Indet_species(Binomial_name)
}
\arguments{
\item{Binomial_name}{}
}
\value{
a character string for the new name of the species
}
\description{
Transforms the name of the undertermined species into "Indet_sp"
}
