% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_variance_comparison.R
\name{variance_comp}
\alias{variance_comp}
\title{Variance comparison}
\usage{
variance_comp(data, nfiles, dataset_name)
}
\arguments{
\item{data}{the whole dataset as a tibble or dataframe}

\item{nfiles}{number of files created for each plot by distlist}

\item{dataset_name}{string of characters - each dataset must begin with a different letter}
}
\value{
a dataframe with the results of the comparison for each species
}
\description{
Variance comparison
}
